Separate WeightedMemberStrategy. Keep Cluster simple.
